create scheduled messages for photos too

+++ main.py
@@ -1,12 +1,13 @@
 from dotenv import load_dotenv
 import os
 from aiogram import Bot, Dispatcher, Router
+from aiogram.utils.media_group import MediaGroupBuilder
 from asyncio import run, sleep
 from app import router
 from loguru import logger
 from multiprocessing import Process
 from utils import convert_data
-from mongo_db import db
+from mongo_db import db, DB
 import time
 import datetime
 
@@ -15,27 +16,20 @@ TOKEN = os.environ.get('TG_BOT_TOKEN')
 
 
 async def send_scheduled_messages(bot: Bot):
+    await sleep(15)
     while True:
         logger.debug('checking for schedule messages')
         msgs = await db.get_scheduled_messages()
-        for msg in msgs:
-            if convert_data(msg['time']) <= datetime.datetime.now():
-                logger.debug('found scheduled message. Sending...')
-                if msg['message'].get('photo', False):
-                    await bot.send_photo(
-                        photo=msg['message']['photo'],
-                        caption=msg['message']['caption'],
-                        chat_id=msg['chatid']
+        if msgs:
+            for msg in msgs:
+                if convert_data(msg['time']) <= datetime.datetime.now():
+                    logger.debug('found scheduled message. Sending...')
+                    await bot.copy_message(
+                        chat_id=msg['chatid'],
+                        from_chat_id=msg['message']['from_chat_id'],
+                        message_id=msg['message']['message_id']
                     )
-                elif msg['message'].get('video', False):
-                    await bot.send_video(
-                        video=msg['message']['video'],
-                        caption=msg['message']['caption'],
-                        chat_id=msg['chatid']
-                    )
-                else:
-                    await bot.send_message(text=msg['message']['text'], chat_id=msg['chatid'])
-                await db.delete_schedule_message_by_id(msg['chatid'])
+                    await db.delete_schedule_message_by_id(msg['chatid'])
         await sleep(60 * 5)  # every 5 minutes
